[BUGFIX] Indexing relations fatals if TCA is not existing
If there is a configuration with SOLR_RELATION conten object
and the configured field does not exist in TCA the indexing
process dies with a fatal error.
The scheduler tasks stays in state "running" without further
notices (f.e. index queue item marked as failed).
The only place to identify the issue is the php error log.

This change adds a graceful behaviour by adding a guard clause.
If the relation to be resolved has no TCA, the field will be
indexed empty.
Resolves: #64789
Releases: master, 3.0

Change-Id: I2b5ab92c3546fcb4faf434dcd02ca618aa5ad136
Reviewed-on: http://review.typo3.org/37090
Reviewed-by: Peter Kraume <peter.kraume@gmx.de>
Tested-by: Peter Kraume <peter.kraume@gmx.de>
Reviewed-by: Ingo Renner <ingo@typo3.org>
Tested-by: Ingo Renner <ingo@typo3.org>