[BUGFIX] Prevent fatal error while indexing
The method getRelatedItems of the content object relations did not
check if the configured local field is defined in the current TCA.

Therefore it can happen that the indexer scheduler task crashes with an
fatal error.

Since commit df04ae1 of tx_news e.g. the default local field
content_elements can be unset.